Renaming model and clean code changes
